Remove home initializer as it take long
